python-backend/projects/digital-art/server/requirements.txt
uvicorn
fastapi
pyjwt
SQLAlchemy
passlib
bcrypt
python-multipart